refactor: naming component
